Move space filling curves and morton code headers
