+++ test/tstDetailsTreeConstruction.cpp
@@ -11,11 +11,11 @@
 #include "ArborXTest_LegacyTree.hpp"
 #include "ArborX_EnableDeviceTypes.hpp" // ARBORX_DEVICE_TYPES
 #include "ArborX_EnableViewComparison.hpp"
-#include <ArborX_MortonCode.hpp> // expandBits, morton32
-#include <ArborX_SortUtils.hpp>  // sortObjects
+#include <ArborX_SortUtils.hpp> // sortObjects
 #include <details/ArborX_Algorithms.hpp>
 #include <details/ArborX_IndexableGetter.hpp>
-#include <details/ArborX_Node.hpp> // ROPE SENTINEL
+#include <details/ArborX_MortonCode.hpp> // expandBits, morton32
+#include <details/ArborX_Node.hpp>       // ROPE SENTINEL
 #include <details/ArborX_TreeConstruction.hpp>
 #include <kokkos_ext/ArborX_KokkosExtStdAlgorithms.hpp>
